Processing extensions prep-work
