+++ src/core/streaming.h
@@ -101,6 +101,7 @@ namespace librealsense
         virtual void acquire() = 0;
         virtual void release() = 0;
         virtual frame_interface* publish(std::shared_ptr<archive_interface> new_owner) = 0;
+        virtual void unpublish() = 0;
         virtual void attach_continuation(frame_continuation&& continuation) = 0;
         virtual void disable_continuation() = 0;
 
@@ -119,7 +120,7 @@ namespace librealsense
         virtual ~frame_interface() = default;
     };
 
-    struct frame_holder
+    struct EXTENSION_API frame_holder
     {
         frame_interface* frame;
